Move code to correct file.
